Updated get_log_location
Updated get_log_location function to return the location of the log file associated with the current PID - if it has already been assigned one. Otherwise, the function create a new log file and assign it to the PID. The code for assigning a new log file has been separated into its own method.