Updated PID database format and handling.
The PID database file now also holds a list of all of the plugins required by the PID's current loaded workflow. Allowing this information to be shared between different instances of the MAP Client. A number of PID database related methods have been added to simplify access to this file.