Update PluginFinderDialog to use remote database.
Since the MAP-plugin database has been moved online, several of the functions related to the generation and visualization of this data have been updated or replaced accordingly:

The lists of plugin_organizations and plugin_repositories are no longer hard coded, but are now part of the remote plugin database instead. This is done with the intention that MAP-Client users will eventually be able to add their own MAP-plugin URLs to the remote database.

The function used to authenticate GitHub personal access tokens has been updated slightly to use a QInputDialog to get user input if needed. All functions that send requests to the GitHub API are now being called by a wrapper function to ensure that the user will be authenticated if the GitHub API rate limit is exceeded.

A number of functions have been added to support local caching of the remote plugin database.

The write_step_database function used by the database generation script has been temporarily disabled. The idea is that in the future most of the generation script will be handled by GitHub actions to help automate the process of updating the database. Because of this, I have chosen to leave the implementation of this function until the addition of these GitHub actions.