Add version checking to PluginFinder tool.
The map-plugin-database has been updated to reference the latest version of each MAP-Client plugin. This commit updates the PluginFinderDialog to use these version numbers to check if the plugins that the user has installed are up to date. The user's current plugin version is displayed to the left of the download button of the corresponding plugin. A downwards arrow (▼) is shown at the end of the version number if the user's plugin is outdated. The download button is disabled only if the user's plugin version matches the most recent version of that plugin.