Relocate PluginFinder visualization code.
Previously, a number of the classes and functions being used to visualize the PluginFinder database were located in the MAP-Client "utils" directory. This was initially due to the fact the some of these classes and functions were also being used in the generation of the MAP plugin-database. Now that the generation of the database has been moved to a GitHub action in the map-plugin-database repository, we can move the remaining visualization code into the directory of the PluginFinder tool.