Update PMR2 DvcsCmd usage
All PMR2 BaseDvcsCmd subclasses now output the Git command return-code in addition to the stdout and stderr streams. This can be used to more accurately filter error messages passed to the MAP-Client from the pmr2.wfctrl package.

A return code check has been added to the PMRTool.pushToRemote method to ensure that the stderr stream is only logged as an error if the return code is non-zero.