Add gitInclude method to WorkflowStepMountPoint
This commit adds two methods to the WorkflowStepMountPoint: gitInclude and createGitIgnore. These methods allow a plugin implementation to specify what files in the step's workflow sub-directory should be included in version control.

The gitInclude method can be overridden by a plugin implementation to specify a list of files or file types that should be included in the version control of the workflow.

The createGitIgnore method is called by the MAP-Client to create a .gitignore file for each of the workflow steps in the workflow directory. This will ignore all files in the step's workflow sub-directory except those specified by the plugin's gitInclude method.