Implement merge methods for QUndoCommand classes
This commit implements mergeWith as well as some supporting methods for the CommandMove class.

It also implements a merge_macros method in the WorkflowGraphicsView class so that we can explicitly merge two QUndoStack macros. This implementation is more of a temporary solution - ideally, we would implement a custom macro class with it's own mergeWith method and let the existing QUndoStack merge functionality handle the rest, but unfortunately explicit macros don't appear to be working with the latest PySide API. The overhead is relatively minimal: having to call the merge method manually and delete the merge macro command when successful - where these calls would otherwise be handled automatically.