Add getIcon and getIcon attributes to MAPPlugin
The MAPPlugin class was originally created to mimic the WorkflowStepMountPoint which would allow us to use the existing addStep function to add "plugin-like" objects to the Plugin Finder GUI. Since there have been updates to the WorkflowStepMountPoint and the addStep method since the PR was submitted, these changes are required to maintain the existing functionality.

Obviously this current implementation isn't very robust. When I get a chance I will look at updating the way the Plugin Finder populates the QListView so that it doesn't rely on the addStep method.